v0.2
Created the GenerateMenu function, which reads in a dictionary of menu
items, and turns them into DirectoryObjects and returns them in a
ObjectContainer. Used this function everywhere I could to cut down on
repetition and lines of code. Created a language file to hold some
string constants. But most importantly, I found some artwork to use as
the background... And I couldn't decide which one to use so I have 2
alternates sitting in the Resources folder. In v0.3 I plan to create
preferences for artwork (and other options), as well as fix up the
MemberMenu function to filter out empty directories.