v0.3
Updated the service to not only select the best possible resolution, but
to actually return all available resolutions. This involves using the
deferred decorator (which is apparently evil, but it seems to work for
me... I've only tested on Roku/Android clients, and I've read threads
from 2013ish that say Plex Home Theatre doesn't work well with it).