Created Video Menu
I have changed the channel so that instead of jumping right to the Video
Clip Object, you are sent to another Directory Object which will contain
the Video Clip Object, as well as related videos, playlists the video is
in, and porn stars in the video. I have not yet added the new features,
just "laid the foundation" by creating the new menu and adding the Video
Clip Object to it. Will be adding new functionality shortly.