Added actions
Added a menu that shows time stamps for actions (e.g. different
positions). I made it using recursive DirectoryObjects because I didn't
really know how else to display information like that. So you can
technically click through that menu infinitely and never get anywhere,
but hopefully after a few clicks you will realize what is happening and
start backing out. If Plex was able to start a video from a certain time
that would be sweet, because it would be trivial to modify this to add a
VideoClipObject with a time offset if that existed. If it does, I don't
know how to do it.