Added additional Member menu items
You can now view a Member's subscribers, Member subscriptions, and
friends. Also re-organized the menu a little bit to bring the content
closer to the top. One thing to note is that while the xPath to check
for a Member's subscribers/Member subscriptions/friends is fairly
accurate, there are times when it is inaccurate. I will be addressing
this in a forthcoming commit where I will add the option to get accurate
results for those menu options (at the expense of 1 extra HTTP request
each). It will be configurable in the Preferences if you don't want to
create any extra HTTP requests