Begun implementing Shared Code
I want to separate the data and logic in my code more. I want the
SharedCodeService to act as a model, and __init__.py (and other files in
that directory) to work as a controller. This commit is the first step
to that. I have moved all of the more basic xPath logic into the
SharedCodeService, removing xPath entirely from several of the
"controllers". There is still much more work to be done on this front,
v1.2 will largely be a code rewrite with only a couple new features.
With any luck though, the code rewrite will allow other features to be
added more easily in the future (and the code should be a bit clearer,
hopefully)