Simplified XPath selectors for video listing, added error routine
The new layout of the website broke the plugin because the XPath selectors in function GetVideos were too specific. I have simplified those so that they should work even with newer versions of the website. I have also added error-handling to said function so that a single malformed video-section does not make the whole plugin crash.